backend/requirements.txt
Flask
Flask-CORS
requests
numpy
python-dotenv 